+++ api-internal/src/main/java/at/pavlov/internal/utils/NumberTricks.java
@@ -1,4 +1,4 @@
-package at.pavlov.internal;
+package at.pavlov.internal.utils;
 
 import lombok.experimental.UtilityClass;
